projects/210306G-Autonomous-Systems_Motion-Prediction/requirements.txt
# Core ML Libraries
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning (uncomment as needed)
# torch>=1.10.0
# torchvision>=0.11.0
# tensorflow>=2.8.0
# keras>=2.8.0

# NLP Libraries (uncomment as needed)
# nltk>=3.7.0
# spacy>=3.4.0
# transformers>=4.15.0

# Computer Vision (uncomment as needed)
# opencv-python>=4.5.0
# pillow>=8.3.0

# Utilities
jupyter>=1.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0

# Add your specific requirements below

# Python environment
python==3.7

# Core PyTorch (CUDA 11.0 build)
torch==1.7.1+cu110
torchvision==0.8.2+cu110
torch-geometric==1.6.3
torch-sparse==0.6.9
torch-scatter==2.0.6

# Chemistry and utilities
rdkit==2020.09.1.0
PyYAML
tensorboard

# Optional (for mixed-precision training)
nvidia-apex
